Refactor CasadiSolver to OpenMP and Cuda classes
