+++ pybamm/solvers/c_solvers/idaklu.cpp
@@ -11,7 +11,7 @@
 
 Function generate_function(const std::string &data)
 {
-  return Function::deserialize(data);
+    return Function::deserialize(data);
 }
 
 namespace py = pybind11;
@@ -20,44 +20,74 @@ PYBIND11_MAKE_OPAQUE(std::vector<np_array>);
 
 PYBIND11_MODULE(idaklu, m)
 {
-  m.doc() = "sundials solvers"; // optional module docstring
+    m.doc() = "sundials solvers"; // optional module docstring
 
-  py::bind_vector<std::vector<np_array>>(m, "VectorNdArray");
+    py::bind_vector<std::vector<np_array>>(m, "VectorNdArray");
 
-  m.def("solve_python", &solve_python,
-        "The solve function for python evaluators", py::arg("t"), py::arg("y0"),
-        py::arg("yp0"), py::arg("res"), py::arg("jac"), py::arg("sens"),
-        py::arg("get_jac_data"), py::arg("get_jac_row_vals"),
-        py::arg("get_jac_col_ptr"), py::arg("nnz"), py::arg("events"),
-        py::arg("number_of_events"), py::arg("use_jacobian"),
-        py::arg("rhs_alg_id"), py::arg("atol"), py::arg("rtol"),
-        py::arg("inputs"), py::arg("number_of_sensitivity_parameters"),
-        py::return_value_policy::take_ownership);
+    m.def("solve_python", &solve_python,
+          "The solve function for python evaluators",
+          py::arg("t"),
+          py::arg("y0"),
+          py::arg("yp0"),
+          py::arg("res"),
+          py::arg("jac"),
+          py::arg("sens"),
+          py::arg("get_jac_data"),
+          py::arg("get_jac_row_vals"),
+          py::arg("get_jac_col_ptr"),
+          py::arg("nnz"),
+          py::arg("events"),
+          py::arg("number_of_events"),
+          py::arg("use_jacobian"),
+          py::arg("rhs_alg_id"),
+          py::arg("atol"),
+          py::arg("rtol"),
+          py::arg("inputs"),
+          py::arg("number_of_sensitivity_parameters"),
+          py::return_value_policy::take_ownership);
 
-  py::class_<CasadiSolver>(m, "CasadiSolver")
-      .def("solve", &CasadiSolver::solve, "perform a solve", py::arg("t"),
-           py::arg("y0"), py::arg("yp0"), py::arg("inputs"),
-           py::return_value_policy::take_ownership);
+    py::class_<CasadiSolver>(m, "CasadiSolver")
+    .def("solve", &CasadiSolver::solve,
+         "perform a solve",
+         py::arg("t"),
+         py::arg("y0"),
+         py::arg("yp0"),
+         py::arg("inputs"),
+         py::return_value_policy::take_ownership);
 
-  m.def("create_casadi_solver", &create_casadi_solver,
-        "Create a casadi idaklu solver object", py::arg("number_of_states"),
-        py::arg("number_of_parameters"), py::arg("rhs_alg"),
-        py::arg("jac_times_cjmass"), py::arg("jac_times_cjmass_colptrs"),
-        py::arg("jac_times_cjmass_rowvals"), py::arg("jac_times_cjmass_nnz"),
-        py::arg("jac_bandwidth_lower"), py::arg("jac_bandwidth_upper"),
-        py::arg("jac_action"), py::arg("mass_action"), py::arg("sens"),
-        py::arg("events"), py::arg("number_of_events"), py::arg("rhs_alg_id"),
-        py::arg("atol"), py::arg("rtol"), py::arg("inputs"), py::arg("options"),
-        py::return_value_policy::take_ownership);
+    m.def("create_casadi_solver", &create_casadi_solver,
+          "Create a casadi idaklu solver object",
+          py::arg("number_of_states"),
+          py::arg("number_of_parameters"),
+          py::arg("rhs_alg"),
+          py::arg("jac_times_cjmass"),
+          py::arg("jac_times_cjmass_colptrs"),
+          py::arg("jac_times_cjmass_rowvals"),
+          py::arg("jac_times_cjmass_nnz"),
+          py::arg("jac_bandwidth_lower"),
+          py::arg("jac_bandwidth_upper"),
+          py::arg("jac_action"),
+          py::arg("mass_action"),
+          py::arg("sens"),
+          py::arg("events"),
+          py::arg("number_of_events"),
+          py::arg("rhs_alg_id"),
+          py::arg("atol"),
+          py::arg("rtol"),
+          py::arg("inputs"),
+          py::arg("options"),
+          py::return_value_policy::take_ownership);
 
-  m.def("generate_function", &generate_function, "Generate a casadi function",
-        py::arg("string"), py::return_value_policy::take_ownership);
+    m.def("generate_function", &generate_function,
+          "Generate a casadi function",
+          py::arg("string"),
+          py::return_value_policy::take_ownership);
 
-  py::class_<Function>(m, "Function");
+    py::class_<Function>(m, "Function");
 
-  py::class_<Solution>(m, "solution")
-      .def_readwrite("t", &Solution::t)
-      .def_readwrite("y", &Solution::y)
-      .def_readwrite("yS", &Solution::yS)
-      .def_readwrite("flag", &Solution::flag);
+    py::class_<Solution>(m, "solution")
+    .def_readwrite("t", &Solution::t)
+    .def_readwrite("y", &Solution::y)
+    .def_readwrite("yS", &Solution::yS)
+    .def_readwrite("flag", &Solution::flag);
 }
