+++ pybamm/solvers/c_solvers/idaklu/CasadiSolver_OpenMP.hpp
@@ -1,0 +1,230 @@
+#ifndef PYBAMM_IDAKLU_CASADI_SOLVER_OPENMP_HPP
+#define PYBAMM_IDAKLU_CASADI_SOLVER_OPENMP_HPP
+
+#include "casadi_solver.hpp"
+
+class CasadiSolver_Dense : public CasadiSolver {
+public:
+  CasadiSolver_Dense(
+    np_array atol_np,
+    double rel_tol,
+    np_array rhs_alg_id,
+    int number_of_parameters,
+    int number_of_events,
+    int jac_times_cjmass_nnz,
+    int jac_bandwidth_lower,
+    int jac_bandwidth_upper,
+    std::unique_ptr<CasadiFunctions> functions,
+    const Options& options
+  ) :
+    CasadiSolver(
+      atol_np,
+      rel_tol,
+      rhs_alg_id,
+      number_of_parameters,
+      number_of_events,
+      jac_times_cjmass_nnz,
+      jac_bandwidth_lower,
+      jac_bandwidth_upper,
+      std::move(functions),
+      options
+    )
+  {
+    Initialize();
+  }
+  void SetLinearSolver() override;
+};
+
+class CasadiSolver_KLU : public CasadiSolver {
+public:
+  CasadiSolver_KLU(
+    np_array atol_np,
+    double rel_tol,
+    np_array rhs_alg_id,
+    int number_of_parameters,
+    int number_of_events,
+    int jac_times_cjmass_nnz,
+    int jac_bandwidth_lower,
+    int jac_bandwidth_upper,
+    std::unique_ptr<CasadiFunctions> functions,
+    const Options& options
+  ) :
+    CasadiSolver(
+      atol_np,
+      rel_tol,
+      rhs_alg_id,
+      number_of_parameters,
+      number_of_events,
+      jac_times_cjmass_nnz,
+      jac_bandwidth_lower,
+      jac_bandwidth_upper,
+      std::move(functions),
+      options
+    )
+  {
+    Initialize();
+  }
+  void SetLinearSolver() override;
+};
+
+class CasadiSolver_Band : public CasadiSolver {
+public:
+  CasadiSolver_Band(
+    np_array atol_np,
+    double rel_tol,
+    np_array rhs_alg_id,
+    int number_of_parameters,
+    int number_of_events,
+    int jac_times_cjmass_nnz,
+    int jac_bandwidth_lower,
+    int jac_bandwidth_upper,
+    std::unique_ptr<CasadiFunctions> functions,
+    const Options& options
+  ) :
+    CasadiSolver(
+      atol_np,
+      rel_tol,
+      rhs_alg_id,
+      number_of_parameters,
+      number_of_events,
+      jac_times_cjmass_nnz,
+      jac_bandwidth_lower,
+      jac_bandwidth_upper,
+      std::move(functions),
+      options
+    )
+  {
+    Initialize();
+  }
+  void SetLinearSolver() override;
+};
+
+class CasadiSolver_SPBCGS : public CasadiSolver {
+public:
+  CasadiSolver_SPBCGS(
+    np_array atol_np,
+    double rel_tol,
+    np_array rhs_alg_id,
+    int number_of_parameters,
+    int number_of_events,
+    int jac_times_cjmass_nnz,
+    int jac_bandwidth_lower,
+    int jac_bandwidth_upper,
+    std::unique_ptr<CasadiFunctions> functions,
+    const Options& options
+  ) :
+    CasadiSolver(
+      atol_np,
+      rel_tol,
+      rhs_alg_id,
+      number_of_parameters,
+      number_of_events,
+      jac_times_cjmass_nnz,
+      jac_bandwidth_lower,
+      jac_bandwidth_upper,
+      std::move(functions),
+      options
+    )
+  {
+    Initialize();
+  }
+  void SetLinearSolver() override;
+};
+
+class CasadiSolver_SPFGMR : public CasadiSolver {
+public:
+  CasadiSolver_SPFGMR(
+    np_array atol_np,
+    double rel_tol,
+    np_array rhs_alg_id,
+    int number_of_parameters,
+    int number_of_events,
+    int jac_times_cjmass_nnz,
+    int jac_bandwidth_lower,
+    int jac_bandwidth_upper,
+    std::unique_ptr<CasadiFunctions> functions,
+    const Options& options
+  ) :
+    CasadiSolver(
+      atol_np,
+      rel_tol,
+      rhs_alg_id,
+      number_of_parameters,
+      number_of_events,
+      jac_times_cjmass_nnz,
+      jac_bandwidth_lower,
+      jac_bandwidth_upper,
+      std::move(functions),
+      options
+    )
+  {
+    Initialize();
+  }
+  void SetLinearSolver() override;
+};
+
+class CasadiSolver_SPGMR : public CasadiSolver {
+public:
+  CasadiSolver_SPGMR(
+    np_array atol_np,
+    double rel_tol,
+    np_array rhs_alg_id,
+    int number_of_parameters,
+    int number_of_events,
+    int jac_times_cjmass_nnz,
+    int jac_bandwidth_lower,
+    int jac_bandwidth_upper,
+    std::unique_ptr<CasadiFunctions> functions,
+    const Options& options
+  ) :
+    CasadiSolver(
+      atol_np,
+      rel_tol,
+      rhs_alg_id,
+      number_of_parameters,
+      number_of_events,
+      jac_times_cjmass_nnz,
+      jac_bandwidth_lower,
+      jac_bandwidth_upper,
+      std::move(functions),
+      options
+    )
+  {
+    Initialize();
+  }
+  void SetLinearSolver() override;
+};
+
+class CasadiSolver_SPTFQMR : public CasadiSolver {
+public:
+  CasadiSolver_SPTFQMR(
+    np_array atol_np,
+    double rel_tol,
+    np_array rhs_alg_id,
+    int number_of_parameters,
+    int number_of_events,
+    int jac_times_cjmass_nnz,
+    int jac_bandwidth_lower,
+    int jac_bandwidth_upper,
+    std::unique_ptr<CasadiFunctions> functions,
+    const Options& options
+  ) :
+    CasadiSolver(
+      atol_np,
+      rel_tol,
+      rhs_alg_id,
+      number_of_parameters,
+      number_of_events,
+      jac_times_cjmass_nnz,
+      jac_bandwidth_lower,
+      jac_bandwidth_upper,
+      std::move(functions),
+      options
+    )
+  {
+    Initialize();
+  }
+  void SetLinearSolver() override;
+};
+
+#endif // PYBAMM_IDAKLU_CASADI_SOLVER_OPENMP_HPP
