+++ pybamm/solvers/c_solvers/idaklu/casadi_solver.cpp
@@ -1,489 +1,195 @@
 #include "casadi_solver.hpp"
+#include "CasadiSolver.hpp"
+#include "CasadiSolver_OpenMP.hpp"
+#include "CasadiSolver_Cuda.hpp"
 #include "casadi_sundials_functions.hpp"
 #include "common.hpp"
 #include <idas/idas.h>
 #include <memory>
 
+#define CUDA 1
+#ifdef CUDA
+  #include "CasadiSolver_Cuda.hpp"
+#endif
 
-CasadiSolver *
-create_casadi_solver(int number_of_states, int number_of_parameters,
-                     const Function &rhs_alg, const Function &jac_times_cjmass,
-                     const np_array_int &jac_times_cjmass_colptrs,
-                     const np_array_int &jac_times_cjmass_rowvals,
-                     const int jac_times_cjmass_nnz, 
-                     const int jac_bandwidth_lower, const int jac_bandwidth_upper, 
-                     const Function &jac_action,
-                     const Function &mass_action, const Function &sens,
-                     const Function &events, const int number_of_events,
-                     np_array rhs_alg_id, np_array atol_np, double rel_tol,
-                     int inputs_length, py::dict options)
-{
+CasadiSolver *create_casadi_solver(
+    int number_of_states,
+    int number_of_parameters,
+    const Function &rhs_alg,
+    const Function &jac_times_cjmass,
+    const np_array_int &jac_times_cjmass_colptrs,
+    const np_array_int &jac_times_cjmass_rowvals,
+    const int jac_times_cjmass_nnz,
+    const int jac_bandwidth_lower,
+    const int jac_bandwidth_upper,
+    const Function &jac_action,
+    const Function &mass_action,
+    const Function &sens,
+    const Function &events,
+    const int number_of_events,
+    np_array rhs_alg_id,
+    np_array atol_np,
+    double rel_tol,
+    int inputs_length,
+    py::dict options
+) {
   auto options_cpp = Options(options);
   auto functions = std::make_unique<CasadiFunctions>(
-      rhs_alg, jac_times_cjmass, jac_times_cjmass_nnz, jac_bandwidth_lower, jac_bandwidth_upper,  jac_times_cjmass_rowvals,
-      jac_times_cjmass_colptrs, inputs_length, jac_action, mass_action, sens,
-      events, number_of_states, number_of_events, number_of_parameters,
-      options_cpp);
+    rhs_alg,
+    jac_times_cjmass,
+    jac_times_cjmass_nnz,
+    jac_bandwidth_lower,
+    jac_bandwidth_upper,
+    jac_times_cjmass_rowvals,
+    jac_times_cjmass_colptrs,
+    inputs_length,
+    jac_action,
+    mass_action,
+    sens,
+    events,
+    number_of_states,
+    number_of_events,
+    number_of_parameters,
+    options_cpp
+  );
 
-  return new CasadiSolver(atol_np, rel_tol, rhs_alg_id, number_of_parameters,
-                          number_of_events, jac_times_cjmass_nnz, 
-                          jac_bandwidth_lower, jac_bandwidth_upper,
-                          std::move(functions), options_cpp);
-}
-
-CasadiSolver::CasadiSolver(np_array atol_np, double rel_tol,
-                           np_array rhs_alg_id, int number_of_parameters,
-                           int number_of_events, int jac_times_cjmass_nnz,
-                           int jac_bandwidth_lower, int jac_bandwidth_upper,
-                           std::unique_ptr<CasadiFunctions> functions_arg,
-                           const Options &options)
-    : number_of_states(atol_np.request().size),
-      number_of_parameters(number_of_parameters),
-      number_of_events(number_of_events),
-      jac_times_cjmass_nnz(jac_times_cjmass_nnz),
-      functions(std::move(functions_arg)), options(options)
-{
-  DEBUG("CasadiSolver::CasadiSolver");
-  auto atol = atol_np.unchecked<1>();
-
-  // create context (not supported by older sundial versions)
-  SUNContext_Create(NULL, &sunctx);
+  CasadiSolver *casadiSolver = nullptr;
   
-  // allocate memory for solver
-  ida_mem = IDACreate(sunctx);
-
-  // allocate vectors
-#ifdef CUDA
-  yy = N_VNew_Cuda(number_of_states, sunctx);
-  yp = N_VNew_Cuda(number_of_states, sunctx);
-  avtol = N_VNew_Cuda(number_of_states, sunctx);
-  id = N_VNew_Cuda(number_of_states, sunctx);
-#else
-  int num_threads = options.num_threads;
-  yy = N_VNew_OpenMP(number_of_states, num_threads, sunctx);
-  yp = N_VNew_OpenMP(number_of_states, num_threads, sunctx);
-  avtol = N_VNew_OpenMP(number_of_states, num_threads, sunctx);
-  id = N_VNew_OpenMP(number_of_states, num_threads, sunctx);
-#endif
-
-  if (number_of_parameters > 0)
+  // Instantiate solver class
+  if (options_cpp.linear_solver == "SUNLinSol_Dense")
   {
-    yyS = N_VCloneVectorArray(number_of_parameters, yy);
-    ypS = N_VCloneVectorArray(number_of_parameters, yp);
+      DEBUG("\tsetting SUNLinSol_Dense linear solver");
+      casadiSolver = new CasadiSolver_Dense(
+          atol_np,
+          rel_tol,
+          rhs_alg_id,
+          number_of_parameters,
+          number_of_events,
+          jac_times_cjmass_nnz,
+          jac_bandwidth_lower,
+          jac_bandwidth_upper,
+          std::move(functions),
+          options_cpp
+       );
   }
-
-  // set initial value
-  realtype *atval = N_VGetArrayPointer(avtol);
-  for (int i = 0; i < number_of_states; i++)
+  else if (options_cpp.linear_solver == "SUNLinSol_KLU")
   {
-    atval[i] = atol[i];
+      DEBUG("\tsetting SUNLinSol_KLU linear solver");
+      casadiSolver = new CasadiSolver_KLU(
+          atol_np,
+          rel_tol,
+          rhs_alg_id,
+          number_of_parameters,
+          number_of_events,
+          jac_times_cjmass_nnz,
+          jac_bandwidth_lower,
+          jac_bandwidth_upper,
+          std::move(functions),
+          options_cpp
+       );
   }
-
-  for (int is = 0; is < number_of_parameters; is++)
+  else if (options_cpp.linear_solver == "SUNLinSol_Band")
   {
-    N_VConst(RCONST(0.0), yyS[is]);
-    N_VConst(RCONST(0.0), ypS[is]);
+      DEBUG("\tsetting SUNLinSol_Band linear solver");
+      casadiSolver = new CasadiSolver_Band(
+          atol_np,
+          rel_tol,
+          rhs_alg_id,
+          number_of_parameters,
+          number_of_events,
+          jac_times_cjmass_nnz,
+          jac_bandwidth_lower,
+          jac_bandwidth_upper,
+          std::move(functions),
+          options_cpp
+       );
   }
-
-  // initialise solver
-
-  IDAInit(ida_mem, residual_casadi, 0, yy, yp);
-
-  // set tolerances
-  rtol = RCONST(rel_tol);
-
-  IDASVtolerances(ida_mem, rtol, avtol);
-
-  // set events
-  IDARootInit(ida_mem, number_of_events, events_casadi);
-
-  void *user_data = functions.get();
-  IDASetUserData(ida_mem, user_data);
-
-
-  printf("1\n");
-
-  // setup CUDA handles
-#ifdef CUDA
-  cusparseStatus_t cusp_status = cusparseCreate(&cusp);
-  if (cusp_status != CUSPARSE_STATUS_SUCCESS) {
-    printf("ERROR: could not create cuSPARSE handle\n");
-  }
-  cusolverStatus_t cusol_status = cusolverSpCreate(&cusol);
-  if (cusol_status != CUSOLVER_STATUS_SUCCESS) {
-    printf("ERROR: could not create cuSOLVER handle\n");
-  }
-#endif
-
-  // set matrix
-  if (options.jacobian == "sparse")
+  else if (options_cpp.linear_solver == "SUNLinSol_SPBCGS")
   {
-    DEBUG("\tsetting sparse matrix");
-#ifdef CUDA
-    if (options.linear_solver == "SUNLinSol_cuSolverSp_batchQR") {
-      J = SUNMatrix_cuSparse_NewCSR(number_of_states, number_of_states,
-                        jac_times_cjmass_nnz, cusp, sunctx);
-    }
-    else
-    {
-      J = SUNSparseMatrix(number_of_states, number_of_states,
-                        jac_times_cjmass_nnz, CSC_MAT, sunctx);
-    }
-#else
-    J = SUNSparseMatrix(number_of_states, number_of_states,
-                      jac_times_cjmass_nnz, CSC_MAT, sunctx);
-#endif
+      DEBUG("\tsetting SUNLinSol_SPBCGS_linear solver");
+      casadiSolver = new CasadiSolver_SPBCGS(
+          atol_np,
+          rel_tol,
+          rhs_alg_id,
+          number_of_parameters,
+          number_of_events,
+          jac_times_cjmass_nnz,
+          jac_bandwidth_lower,
+          jac_bandwidth_upper,
+          std::move(functions),
+          options_cpp
+       );
   }
-  else if (options.jacobian == "banded") {
-    DEBUG("\tsetting banded matrix");
-      J = SUNBandMatrix(number_of_states, jac_bandwidth_upper, jac_bandwidth_lower, sunctx);
-  } else if (options.jacobian == "dense" || options.jacobian == "none")
+  else if (options_cpp.linear_solver == "SUNLinSol_SPFGMR")
   {
-    DEBUG("\tsetting dense matrix");
-    J = SUNDenseMatrix(number_of_states, number_of_states, sunctx);
+      DEBUG("\tsetting SUNLinSol_SPFGMR_linear solver");
+      casadiSolver = new CasadiSolver_SPFGMR(
+          atol_np,
+          rel_tol,
+          rhs_alg_id,
+          number_of_parameters,
+          number_of_events,
+          jac_times_cjmass_nnz,
+          jac_bandwidth_lower,
+          jac_bandwidth_upper,
+          std::move(functions),
+          options_cpp
+       );
   }
-  else if (options.jacobian == "matrix-free")
+  else if (options_cpp.linear_solver == "SUNLinSol_SPGMR")
   {
-    DEBUG("\tsetting matrix-free");
-    J = NULL;
+      DEBUG("\tsetting SUNLinSol_SPGMR solver");
+      casadiSolver = new CasadiSolver_SPGMR(
+          atol_np,
+          rel_tol,
+          rhs_alg_id,
+          number_of_parameters,
+          number_of_events,
+          jac_times_cjmass_nnz,
+          jac_bandwidth_lower,
+          jac_bandwidth_upper,
+          std::move(functions),
+          options_cpp
+       );
   }
-  printf("2\n");
-
-#if SUNDIALS_VERSION_MAJOR >= 6
-  int precon_type = SUN_PREC_NONE;
-  if (options.preconditioner != "none") {
-    precon_type = SUN_PREC_LEFT;
-  }
-#else
-  int precon_type = PREC_NONE;
-  if (options.preconditioner != "none") {
-    precon_type = PREC_LEFT;
-  }
-#endif
-
-  // set linear solver
-  if (options.linear_solver == "SUNLinSol_Dense")
+  else if (options_cpp.linear_solver == "SUNLinSol_SPTFQMR")
   {
-    DEBUG("\tsetting SUNLinSol_Dense linear solver");
-    LS = SUNLinSol_Dense(yy, J, sunctx);
-  }
-  else if (options.linear_solver == "SUNLinSol_KLU")
-  {
-    DEBUG("\tsetting SUNLinSol_KLU linear solver");
-    LS = SUNLinSol_KLU(yy, J, sunctx);
-  }
-  else if (options.linear_solver == "SUNLinSol_Band")
-  {
-    DEBUG("\tsetting SUNLinSol_Band linear solver");  
-    LS = SUNLinSol_Band(yy, J, sunctx);
-  }
-  else if (options.linear_solver == "SUNLinSol_SPBCGS")
-  {
-    DEBUG("\tsetting SUNLinSol_SPBCGS_linear solver");
-    LS = SUNLinSol_SPBCGS(yy, precon_type, options.linsol_max_iterations,
-                          sunctx);
-  }
-  else if (options.linear_solver == "SUNLinSol_SPFGMR")
-  {
-    DEBUG("\tsetting SUNLinSol_SPFGMR_linear solver");
-    LS = SUNLinSol_SPFGMR(yy, precon_type, options.linsol_max_iterations,
-                          sunctx);
-  }
-  else if (options.linear_solver == "SUNLinSol_SPGMR")
-  {
-    DEBUG("\tsetting SUNLinSol_SPGMR solver");
-    LS = SUNLinSol_SPGMR(yy, precon_type, options.linsol_max_iterations,
-                          sunctx);
-  }
-  else if (options.linear_solver == "SUNLinSol_SPTFQMR")
-  {
-    DEBUG("\tsetting SUNLinSol_SPGMR solver");
-    LS = SUNLinSol_SPTFQMR(yy, precon_type, options.linsol_max_iterations,
-                          sunctx);
+      DEBUG("\tsetting SUNLinSol_SPGMR solver");
+      casadiSolver = new CasadiSolver_SPTFQMR(
+          atol_np,
+          rel_tol,
+          rhs_alg_id,
+          number_of_parameters,
+          number_of_events,
+          jac_times_cjmass_nnz,
+          jac_bandwidth_lower,
+          jac_bandwidth_upper,
+          std::move(functions),
+          options_cpp
+       );
   }
 #ifdef CUDA
-  else if (options.linear_solver == "SUNLinSol_cuSolverSp_batchQR")
+  else if (options_cpp.linear_solver == "SUNLinSol_cuSolverSp_batchQR")
   {
-    DEBUG("\tsetting SUNLinSol_cuSolverSp_batchQR solver");
-    LS = SUNLinSol_cuSolverSp_batchQR(yy, J, cusol, sunctx);
+      DEBUG("\tsetting SUNLinSol_cuSolverSp_batchQR solver");
+      casadiSolver = new CasadiSolver_cuSolverSp_batchQR(
+          atol_np,
+          rel_tol,
+          rhs_alg_id,
+          number_of_parameters,
+          number_of_events,
+          jac_times_cjmass_nnz,
+          jac_bandwidth_lower,
+          jac_bandwidth_upper,
+          std::move(functions),
+          options_cpp
+       );
   }
 #endif
-  printf("3\n");
 
-
-
-  IDASetLinearSolver(ida_mem, LS, J);
-
-  if (options.preconditioner != "none")
-  {
-    DEBUG("\tsetting IDADDB preconditioner");
-    // setup preconditioner
-    IDABBDPrecInit(
-        ida_mem, number_of_states, options.precon_half_bandwidth,
-        options.precon_half_bandwidth, options.precon_half_bandwidth_keep,
-        options.precon_half_bandwidth_keep, 0.0, residual_casadi_approx, NULL);
+  if (casadiSolver == nullptr) {
+    throw std::invalid_argument("Unsupported solver requested");
   }
 
-  if (options.jacobian == "matrix-free")
-  {
-    IDASetJacTimes(ida_mem, NULL, jtimes_casadi);
-  }
-  else if (options.jacobian != "none")
-  {
-    IDASetJacFn(ida_mem, jacobian_casadi);
-  }
-
-  if (number_of_parameters > 0)
-  {
-    IDASensInit(ida_mem, number_of_parameters, IDA_SIMULTANEOUS,
-                sensitivities_casadi, yyS, ypS);
-    IDASensEEtolerances(ida_mem);
-  }
-
-  SUNLinSolInitialize(LS);
-
-  auto id_np_val = rhs_alg_id.unchecked<1>();
-  realtype *id_val;
-  id_val = N_VGetArrayPointer(id);
-
-  int ii;
-  for (ii = 0; ii < number_of_states; ii++)
-  {
-    id_val[ii] = id_np_val[ii];
-  }
-
-  IDASetId(ida_mem, id);
-}
-
-CasadiSolver::~CasadiSolver()
-{
-
-  /* Free memory */
-  if (number_of_parameters > 0)
-  {
-    IDASensFree(ida_mem);
-  }
-  SUNLinSolFree(LS);
-  SUNMatDestroy(J);
-  N_VDestroy(avtol);
-  N_VDestroy(yy);
-  N_VDestroy(yp);
-  N_VDestroy(id);
-  if (number_of_parameters > 0)
-  {
-    N_VDestroyVectorArray(yyS, number_of_parameters);
-    N_VDestroyVectorArray(ypS, number_of_parameters);
-  }
-
-#ifdef CUDA
-  cusparseDestroy(cusp);
-  cusolverSpDestroy(cusol);
-#endif
-  IDAFree(&ida_mem);
-  SUNContext_Free(&sunctx);
-}
-
-Solution CasadiSolver::solve(np_array t_np, np_array y0_np, np_array yp0_np,
-                             np_array_dense inputs)
-{
-  DEBUG("CasadiSolver::solve");
-
-  int number_of_timesteps = t_np.request().size;
-  auto t = t_np.unchecked<1>();
-  realtype t0 = RCONST(t(0));
-  auto y0 = y0_np.unchecked<1>();
-  auto yp0 = yp0_np.unchecked<1>();
-
-
-  if (y0.size() != number_of_states + number_of_parameters * number_of_states) {
-    throw std::domain_error(
-      "y0 has wrong size. Expected " + 
-      std::to_string(number_of_states + number_of_parameters * number_of_states) + 
-      " but got " + std::to_string(y0.size()));
-  }
-
-  if (yp0.size() != number_of_states + number_of_parameters * number_of_states) {
-    throw std::domain_error(
-      "yp0 has wrong size. Expected " + 
-      std::to_string(number_of_states + number_of_parameters * number_of_states) + 
-      " but got " + std::to_string(yp0.size()));
-  }
-
-  // set inputs
-  auto p_inputs = inputs.unchecked<2>();
-  for (int i = 0; i < functions->inputs.size(); i++)
-  {
-    functions->inputs[i] = p_inputs(i, 0);
-  }
-
-  // set initial conditions
-  realtype *yval = N_VGetArrayPointer(yy);
-  realtype *ypval = N_VGetArrayPointer(yp);
-  std::vector<realtype *> ySval(number_of_parameters);
-  std::vector<realtype *> ypSval(number_of_parameters);
-  for (int p = 0 ; p < number_of_parameters; p++) {
-    ySval[p] = N_VGetArrayPointer(yyS[p]);
-    ypSval[p] = N_VGetArrayPointer(ypS[p]);
-    for (int i = 0; i < number_of_states; i++) {
-      ySval[p][i] = y0[i + (p + 1) * number_of_states];
-      ypSval[p][i] = yp0[i + (p + 1) * number_of_states];
-    }
-  }
-
-  for (int i = 0; i < number_of_states; i++)
-  {
-    yval[i] = y0[i];
-    ypval[i] = yp0[i];
-  }
-
-  IDAReInit(ida_mem, t0, yy, yp);
-  if (number_of_parameters > 0) {
-    IDASensReInit(ida_mem, IDA_SIMULTANEOUS, yyS, ypS);
-  }
-
-  // calculate consistent initial conditions
-  DEBUG("IDACalcIC");
-  IDACalcIC(ida_mem, IDA_YA_YDP_INIT, t(1));
-  if (number_of_parameters > 0)
-  {
-    IDAGetSens(ida_mem, &t0, yyS);
-  }
-
-  int t_i = 1;
-  realtype tret;
-  realtype t_next;
-  realtype t_final = t(number_of_timesteps - 1);
-
-  // set return vectors
-  realtype *t_return = new realtype[number_of_timesteps];
-  realtype *y_return = new realtype[number_of_timesteps * number_of_states];
-  realtype *yS_return = new realtype[number_of_parameters *
-                                     number_of_timesteps * number_of_states];
-
-  py::capsule free_t_when_done(t_return,
-                               [](void *f)
-                               {
-                                 realtype *vect =
-                                     reinterpret_cast<realtype *>(f);
-                                 delete[] vect;
-                               });
-  py::capsule free_y_when_done(y_return,
-                               [](void *f)
-                               {
-                                 realtype *vect =
-                                     reinterpret_cast<realtype *>(f);
-                                 delete[] vect;
-                               });
-  py::capsule free_yS_when_done(yS_return,
-                                [](void *f)
-                                {
-                                  realtype *vect =
-                                      reinterpret_cast<realtype *>(f);
-                                  delete[] vect;
-                                });
-
-  t_return[0] = t(0);
-  for (int j = 0; j < number_of_states; j++)
-  {
-    y_return[j] = yval[j];
-  }
-  for (int j = 0; j < number_of_parameters; j++)
-  {
-    const int base_index = j * number_of_timesteps * number_of_states;
-    for (int k = 0; k < number_of_states; k++)
-    {
-      yS_return[base_index + k] = ySval[j][k];
-    }
-  }
-
-
-
-  int retval;
-  while (true)
-  {
-    t_next = t(t_i);
-    IDASetStopTime(ida_mem, t_next);
-    DEBUG("IDASolve");
-    retval = IDASolve(ida_mem, t_final, &tret, yy, yp, IDA_NORMAL);
-
-    if (retval == IDA_TSTOP_RETURN || retval == IDA_SUCCESS ||
-        retval == IDA_ROOT_RETURN)
-    {
-      if (number_of_parameters > 0)
-      {
-        IDAGetSens(ida_mem, &tret, yyS);
-      }
-
-      t_return[t_i] = tret;
-      for (int j = 0; j < number_of_states; j++)
-      {
-        y_return[t_i * number_of_states + j] = yval[j];
-      }
-      for (int j = 0; j < number_of_parameters; j++)
-      {
-        const int base_index =
-            j * number_of_timesteps * number_of_states + t_i * number_of_states;
-        for (int k = 0; k < number_of_states; k++)
-        {
-          yS_return[base_index + k] = ySval[j][k];
-        }
-      }
-      t_i += 1;
-      if (retval == IDA_SUCCESS || retval == IDA_ROOT_RETURN)
-      {
-        break;
-      }
-    }
-    else
-    {
-      // failed
-      break;
-    }
-  }
-
-  np_array t_ret = np_array(t_i, &t_return[0], free_t_when_done);
-  np_array y_ret =
-      np_array(t_i * number_of_states, &y_return[0], free_y_when_done);
-  np_array yS_ret = np_array(
-      std::vector<ptrdiff_t>{number_of_parameters, number_of_timesteps, number_of_states},
-      &yS_return[0], free_yS_when_done);
-
-  Solution sol(retval, t_ret, y_ret, yS_ret);
-
-  if (options.print_stats)
-  {
-    long nsteps, nrevals, nlinsetups, netfails;
-    int klast, kcur;
-    realtype hinused, hlast, hcur, tcur;
-
-    IDAGetIntegratorStats(ida_mem, &nsteps, &nrevals, &nlinsetups, &netfails,
-                          &klast, &kcur, &hinused, &hlast, &hcur, &tcur);
-
-    long nniters, nncfails;
-    IDAGetNonlinSolvStats(ida_mem, &nniters, &nncfails);
-
-    long int ngevalsBBDP = 0;
-    if (options.using_iterative_solver)
-    {
-      IDABBDPrecGetNumGfnEvals(ida_mem, &ngevalsBBDP);
-    }
-
-    py::print("Solver Stats:");
-    py::print("\tNumber of steps =", nsteps);
-    py::print("\tNumber of calls to residual function =", nrevals);
-    py::print("\tNumber of calls to residual function in preconditioner =",
-              ngevalsBBDP);
-    py::print("\tNumber of linear solver setup calls =", nlinsetups);
-    py::print("\tNumber of error test failures =", netfails);
-    py::print("\tMethod order used on last step =", klast);
-    py::print("\tMethod order used on next step =", kcur);
-    py::print("\tInitial step size =", hinused);
-    py::print("\tStep size on last step =", hlast);
-    py::print("\tStep size on next step =", hcur);
-    py::print("\tCurrent internal time reached =", tcur);
-    py::print("\tNumber of nonlinear iterations performed =", nniters);
-    py::print("\tNumber of nonlinear convergence failures =", nncfails);
-  }
-
-  return sol;
+  return casadiSolver;
 }
